Just a consistency with other variables.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@44972 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/exe.h
@@ -919,7 +919,7 @@ public:
 		jrd_nod* csb_message;			// Msg for send/receive
 		Format* csb_format;				// Default Format for stream
 		UInt32Bitmap* csb_fields;		// Fields referenced
-		float csb_cardinality;			// Cardinality of relation
+		double csb_cardinality;			// Cardinality of relation
 		jrd_nod* csb_plan;				// user-specified plan for this relation
 		UCHAR* csb_map;					// Stream map for views
 		RecordSource** csb_rsb_ptr;		// point to rsb for nod_stream
